Created the CSS file
This file contains the CSS code and this CSS file is linked to the HTML file in this same repository.  
Steps to run this frontend part of a student dashboard is-
Open the HTML and CSS files given in this repository in a single folder the folder, save the folder and files and then start Live Server. Will be able to see the web page.